Output: select fields you want for listings
Closes #24.

There is one gotcha. When user writes the option mistakenly like
`--fields`, Thor automatically parses it the same way as if user had
written `--fields=fields`, even though the `fields` option is configured
to type `string`, so it should not be treated like a switch imho. Seems
to me more like a bug than a feature, but I don't know how to write a
*clean* wrokaround, so I'd rather keep it this way for now.