Validate config for tasks other than 'help'
CLI now checks that conductor url, username and password were set by the
user and prints an error message if they are not set.

Thor's 'required options' feature cannot be used here, because the
options can be set either on command line or in config file, so the
checking logic has to be custom.

The config is not checked when running help tasks. User should be able
to view usage information of CLI without having to specify e.g.
the URL of Conductor.